Rename class CommittedImages to Images
It's not always for committed images.
